feat(PromptPreview): 폼 입력시 미리보기 화면에 입력 내용 나타나도록 구현
