+++ card-capture/src/app/prompt/page.tsx
@@ -3,14 +3,22 @@
 import PromptTitle from '@/components/prompt/PromptTitle/PromptTitle';
 import NavigationBar from '@/components/common/NavigationBar/NavigationBar';
 import React from 'react';
-import PromptInput, { PromptInputFormType } from '@/components/prompt/PromptInput/PromptInput';
+import PromptInput from '@/components/prompt/PromptInput/PromptInput';
 import PromptPreview from '@/components/prompt/PromptPreview/PromptPreview';
 import { useFieldArray, useForm } from 'react-hook-form';
+
+export type PromptInputFormType = {
+  phrases: { value: string }[];
+  emphasis: { value: string }[];
+  purpose: string;
+  color: string;
+  model: string;
+};
 
 const PromptPage = () => {
   const formMethods = useForm<PromptInputFormType>({
     mode: 'onBlur',
-    defaultValues: { phrases: [{ value: '' }] },
+    defaultValues: { phrases: [{ value: '' }], color: '#000000' },
   });
 
   const phraseFieldArray = useFieldArray({ control: formMethods.control, name: 'phrases' });
@@ -25,7 +33,7 @@ const PromptPage = () => {
         <PromptTitle />
         <div className="flex flex-col items-center justify-center gap-[50px] px-[20px] py-[50px] xs:px-[50px] md:py-[70px] lg:flex-row lg:items-start">
           <PromptInput formMethods={formMethods} fieldArrays={fieldArrays} />
-          <PromptPreview />
+          <PromptPreview formData={formMethods.watch()} />
         </div>
       </div>
     </div>
